labirynth and minimap
Now at the start of the program a maze is randomly generated with a simple stack algorithm. Corridor can be of any width, but at 1 it gets too claustrophobic and hard to navigate. Also a minimap is included, for debugging and orientation purposes - it can scale to any pixel^2 size.